third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbfactory_open11-expected.txt

This is a testharness.js-based test.
PASS IDBFactory.open() - second open's transaction is available to get objectStores 
Harness: the test ran to completion.
